Modifying filetype collision behavior
Changed some of the autocommand so the filetype detection runs after the
extension file name autocommand runs. I also found out that the
BufEnter and BufWinEnter switch order when opening the file from the
file navigation (Netrw).

The collision resolution of setting the global variable is very
"invasive" which the documentation already states. Changing to
`setfiletype` is less invasive. As such the documentation is unchanged
and the user still has full control by setting the extensions associated
with MOOSE input files.

Closes #21